[Webpack:Update] Add two files of webpackConfig to prod and dev
